+++ app/src/main/java/ru/hepolise/volumekeytrackcontrol/util/VibratorUtil.kt
@@ -32,7 +32,7 @@ object VibratorUtil {
         } else {
             this.vibrate(
                 VibrationEffect.createOneShot(
-                    prefs.getVibrationLength(),
+                    prefs.getVibrationLength().toLong(),
                     prefs.getVibrationAmplitude()
                 )
             )
